Update recipe schema for new globs  and `files` (#34)
This adds the following:

```yaml
build:
  files:
    # select files to include based on a glob expression
    - foo/*
```

It also extends the `glob` definition (in `rattler-build` that's the
`GlobVec` type). You can use either a single glob, a list of globs
(treated as `include` globs) or a map with `include` and `exclude`
globs.

```yaml
files: foo/*
# or 
files:
  - foo/*
# or 
files:
  include:
    - foo/*
  exclude:
    - *.txt
```

All three forms are valid in PR
https://github.com/prefix-dev/rattler-build/pull/819

Note: CEP has to be updated as well.